update selector component naming
